finishing detail coin

+++ app/src/main/java/com/catnip/mycoin/data/network/model/response/coin/detail/CurrentPrice.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 
 data class CurrentPrice(
     @SerializedName("usd")
-    var usd: Int? = null
+    var usd: Double? = null
 )
